split CliEvent* into its own file
With some minor doc updates.
